Add support for optional keyid_hash_algorithms in keys
The TUF 1.0 keyid algorithm is `hexdigit(sha256(cjson(public_key)))`.
python-tuf has an extension to the spec where keys in the root metadata
contain a `keyid_hash_algorithms` which can specify alternative algorithms
to sha256. In order to maintain compatibility with python-tuf and go-tuf,
this patch adds the ability to parse metadata that contains this field.
It also generates metadata with `keyid_hash_algorithms=["sha256"]`, since
python-tuf currently has a hard requirement that this field must be
present in the root metadata. However, this field is optional in rust-tuf,
and can parse metadata that doesn't contain the field.

Change-Id: I5f52524abf17c51a5e36f5be2dabdc970c910990