Add Ephemeral::{metadata, targets}
This changes Ephemeral repository to allow access to the underlying metadata
and targets that were written to the repository. This allows for direct
modification, in case users want to directly add metadata to the repository.
This should be safe, because the repositories are considered untrusted, and
still needs to go through verification through the workflow before it is used
by the client.

Change-Id: I66eb6726c7683c08c1a3e8cf30331f711de12125