Remove some cascaded functions, change defaults, docs
Going through the rustdocs, the all the overloaded functions were rather
confusing. Since most of these methods are only used in a few special
places, this patch removes most overloads, and keeps:

* RepoBuilder<Root>::commit
* RepoBuilder<Root>::add_target

Next, rather than using `create_*` and `with_*_builder`, it switches to
`stage_*` and `stage_*_with_builder` to make it more obvious that we haven't
written the metadata yet.

Next, it changes the root metadata to default to consistent_snapshots,
since that avoids many race conditions, and is a safer default.

Finally, it cleans up the docs.